+++ src/main/java/com/coloronme/admin/domain/personalColor/repository/PersonalColorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonalColorRepository extends JpaRepository<PersonalColor, Long> {
+public interface PersonalColorRepository extends JpaRepository<PersonalColor, Integer> {
 }
